+++ src/header/mod.rs
@@ -162,6 +162,9 @@ pub use self::name::{
     X_DNS_PREFETCH_CONTROL,
     X_FRAME_OPTIONS,
     X_XSS_PROTECTION,
+    X_FORWARDED_FOR,
+    X_FORWARDED_PROTO,
+    X_FORWARDED_PORT,
 };
 
 /// Maximum length of a header name